Move to user_book dir
